+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+## [0.5.0]
+
+### Changed
+
+- migrate to embedded-hal 1.0 #22
+- provide our own CountDown trait instead of the one from embedded-hal 0.2.7 #22
+- no longer ignore I2C errors in I2CInterface::wait_ready dfc595ec
+
 ## [0.4.0]
 
 ### Added